+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- validation of the annotation index checks for annotation period outside of audio duration
+
 ### Changed
 
 - ignore_discarded is now the default behaviour of childproject projects, robustness was added to the discard column
+- annotation index validation and annotation importation now checks for the range_offset being greater than duration and oiutputs an error when it is the case
 - projects now check for unicity of the experiment column in children and recording csv, read() fails when not unique
 - zooniverse uploads : uploads extra metadata to zooniverse (name of the audio clip, dataset it belongs to)
 - zooniverse uploads : the subject_set id is stored in the chunk csv file, as the subject_set name (display name) is susceptible to change